spellcheck, and definition improvement
Trying to improve the definition of partial and full cover. If you're behind full cover (up to your neck) you shouldn't be able to crouch and get a bonus. You'll just be completely obscured (wasn't sure how to add that definition, if I needed to). 
If you're behind full cover, you shouldn't normally be able to go prone and gain +6 cover bonus.