c2.0 feats
moved empty the mag to 'active feats'. Made it only apply to guns with semi-auto, so that machine-gun players won't have the option to use it since it's probably worse than their base fire rate/doesn't make sense for them.

Enhanced Neural Interface didn't seem useful as a feat, tried to buff.

Improvise Weapon, note here, i'm trying remove unarmed combat skill from the game in favor of a 'melee combat skill' to cover melee as a whole. Unarmed is nigh worthless unless you are a monk; and especially worthless if you have a sword. I believe it's beneficial to the game to group it together into a larger-covering skill, since a typical melee character won't have to invest tons to be really useful in something that _probably_ doesn't pay off 90% of the time. 
Maybe leaving unarmed in as a skill that gives benefits to grappling and unarmed hit chance (and maybe even unarmed damage) will be good for characters who want extra melee prowess. But making 'melee combat skill' more of a global control for melee would make melee more accessible--and more likely to happen as a whole.

run n' gun, added a stamina cost, moved to active feats

Buffed salty hashtables. I believe the spirit of the rules should be to balance the feats and make them similar in power to one-another. Globally granting a +20% bonus to w/e value the feat adjusts, will certainly not be balanced in many cases.

Buffed steady shot, added stamina cost, and moved to active feats.

Added a lot of requirements: primary stat-gates a lot of feats so it's not one-size fits all.

Myriad of other changes.