Implementing Movement in Microgravity
I decided for the sake of complexity to allow for a lot of GM Discretion for things like angles and collision physics. I've also focused almost entirely on movement, restricting the effects of weapons fire to increased difficulty of Anchoring oneself.

I've also hijacked the 02 slot for nonstandard environmental rules. This is of course subject to change on request.

Remember that this has 'Supplemental Rulebook Material' written all over it.