Balance pass on SMG Expertise
SMGs are strong for short-range room clearing and suppression by virtue of automatic fire, but suffer at range, stopping power. and especially accuracy.
Accordingly, the low level bonuses are to the short range brackets and Reflex modifier, and the skill substitution is to intimidate a target you are suppressing.
The later bonuses are to accuracy in automatic fire, R3 when firing semi-automatically (to help the SMG contribute to longer ranged firefights), and the capstone allows you to spend a reaction to duck behind cover while suppressing if someone else shoots at you. The idea with this capstone is that it mitigates a penalty of the suppression mechanic and solidifies the SMG as a suppression weapon.