Added Pho-Ox
Added the poison Pho-Ox, a Nettle Agent. This is based on the real-life
organic chemical weapon, Phosgene oxime which, fun fact as opposed to VX
gas is abrieviated as CX! This chemical is actually corosive to metal,
but in this ability I've radically exagerated how much damage pouring it
into a robot would do with the conciet that this is an unspecified
 derivative of Phosgene Oxime, not the chemical itself.

This poison is to be unlocked at level 3 as an alternative to Harmasist
Spike.

	modified:   Doctor Procedures.txt