Medicalize Cover Slide
Some functional changes here. 
* Now does not require that you end in a space with cover, to make tracking easier for the DM.
* Now consumes two actions instead of one, to avoid slapping a huge cooldown on it.
* Now gives an evasion bonus to compensate for the above. Useful for dodging readied shots and light suppression when moving between covers.

General note on Active Feats with a cooldown of None: I gave this to a few of them, generally because I'm not concerned about multiple uses per round (in the case of cover slide, it would be really hard to do so at all. The intention is that the Cooldown: None feats still require you to have a unit of medi nanites available and off cooldown., it just doesn't then send them into any remission.