Create 04_07_Technician_Equipment
I don't know why or how I came up with the idea but unless someone stops me, every. single. core. is going to be riffing off of Intel. And I will be cackling the whole time.

Anyway, here's some equipment for Technicians. With stats and costs and everything. It's very shiny.